Working  on Setting and Scatterplot
