+++ src/App.css
@@ -1,10 +1,9 @@
 .App {
   height: 100%;
   width: 100%; 
-  display: grid;
-  grid-gap: 20px 10px; 
+  display: grid; 
   grid-template-rows: 50% 50%; 
-  grid-template-columns: 25% 37.5% 37.5%;
+  grid-template-columns: 15% 42% 42%;
   grid-template-areas:  
   "Settings Up-Wordmap Up-Wordmap" 
   "Settings Down-Left-Treemap Down-Right-Scatter" 
@@ -14,7 +13,7 @@
 @media screen and (max-width: 480px) { 
   .App { 
     display: grid;
-    grid-gap: 20px 10px; 
+    grid-gap: 15px; 
     grid-template-rows: 25%  25%  25%  25%; 
     grid-template-columns: 100% ;
     grid-template-areas:
