+++ src/Components/Settings/settings.css
@@ -1,6 +1,12 @@
 .Settings {
+  margin: 3px;
   grid-area: Settings;
-  background-color: yellow;
+  background-color: lightgray;
+  padding: 1px 3px;
+  border-style: solid;
+  border-width: 1px;
+  border-color: black;
+  border-radius: 10px;
 }
 
 .Settings h1 {
